docker/loki_log_shipper/app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
protobuf==3.19.3
python-dateutil==2.8.2
python-snappy==0.6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
